pyproject.toml
[project]
name = "PROYECTO-M5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]